requirements.txt
flask
werkzeug
numpy
pillow
gunicorn
tensorflow-cpu
